requirements.txt
Flask==2.3.2
requests==2.31.0
Pillow==10.0.0
transformers==4.35.0
torch==2.0.1
